Location in the SNAPverse

List the packages in and around the SNAPverse and their section of the verse.

sv_pkgs()

Value

a data frame.

Details

This function returns a 3-column data frame of all SNAPverse package names, their part of the verse, and whether local git repositories/R projects sharing the same parent directory are found for each. The list includes SNAPverse satellite packages, which are related to and often used in conjunction with, but not a strict part of the SNAPverse.

The other packages are considered core packages of the SNAPverse and fall into of three sectors of the verse: functions, data or apps. The three packages listed as sector packages load different subsets of SNAPverse packages for convenience. snapverse loads functions, data and apps packages. snaplite loads only the functions packages. snapdata loads only the data packages. For most users, the packages most likely needed or of interst are snapfuns, snapapps and any of the data packages.

Examples

sv_pkgs

#> function(){
#>   typenames <- c("sector", "functions", "data", "apps", "satellite")
#>   type <- list(
#>     sector=c("snapverse", "snaplite", "snapdata"),
#>     functions=c("snapfuns", "snapprep", "alfresco"),
#>     data=c("snapclim", "snapfire", "snappoly", "snapmaps", "snapdist"),
#>     apps="snapapps",
#>     satellite=c("rvtable", "apputils", "maputils", "snaputils", "snapmeta", "snapsite")
#> 
#>   )
#>   d <- tibble::data_frame(
#>     pkg=unlist(type), type=rep(typenames, times=sapply(type, length)), local=FALSE
#>   )
#>   x <- sv_local_pkgs()
#>   if(length(x)) d$local[d$pkg %in% x] <- TRUE
#>   d
#> }
#> <environment: namespace:snapmeta>